test: add regression tests for unstable enum option ids
