+++ src/analyzer/fixtures/prop-any.ts
