+++ src/analyzer/fixtures/prop-event-function.ts
