+++ src/analyzer/fixtures/prop-event-handler.ts
