+++ src/analyzer/fixtures/prop-event-method.ts
