+++ src/analyzer/fixtures/prop-generic-event-handler.ts
