+++ src/analyzer/fixtures/string-enum.ts
@@ -1,0 +1,4 @@
+export enum TestStringEnum {
+	One = 'one',
+	Two = 'two'
+}
