+++ src/analyzer/typescript-react-analyzer/property-analyzer/property-creators.ts
@@ -1,0 +1,1 @@
+export * from './create-enum-property';
